requirements.txt
Django>=3.2
djangorestframework>=3.12
djongo>=1.3.6
pymongo>=3.11
drf-yasg>=1.20.0
djangorestframework-simplejwt>=4.8.0
